[GovWayCore, GovWayMonitor, TestSuite] Corretti errori introdotti dal precedente commit per la testsuite che verifica la consegna delle notifiche. Migliorata pagina di dettaglio di una consegna asincrona, nella console di monitoraggio, in modo da riportare come dimensione la label 'n.d.' invece di '-1' se un payload non è presente.

+++ tools/utils/src/org/openspcoop2/utils/resources/FileSystemUtilities.java
@@ -271,11 +271,13 @@ public class FileSystemUtilities {
 
         if (dir.isDirectory()) {
             String[] children = dir.list();
-            for (int i=0; i<children.length; i++) {
-                boolean success = FileSystemUtilities.deleteDir(new File(dir, children[i]));
-                if (!success) {
-                    return false;
-                }
+            if(children!=null && children.length>0) {
+            	for (int i=0; i<children.length; i++) {
+            		boolean success = FileSystemUtilities.deleteDir(new File(dir, children[i]));
+            		if (!success) {
+            			return false;
+            		}
+            	}
             }
         }
